fix(unsubscribe): move POST handler to /u/[token]/confirm to resolve Next.js route conflict
Next.js App Router disallows a page.tsx and route.ts at the same path —
they're treated as parallel pages that resolve to the same route. CI's
production build failed where local pnpm typecheck/lint/boundaries hadn't.

Move:
- apps/web/app/u/[token]/route.ts → apps/web/app/u/[token]/confirm/route.ts
- import path adjusted (./_lib → ../_lib)
- Form `action` in unsubscribe-page-view.tsx now POSTs to /confirm
- Test import + URL assertions updated to match

The semantic flow is unchanged: GET renders the pending confirmation page,
form POST to /u/[token]/confirm records the scope-specific opt-out and
303-redirects to /u/[token]?confirmed=1, which renders the success state.

The /all/route.ts handler (POST for all-scope opt-out) is unaffected.

Process note: should have run `pnpm build` locally before pushing.
Adding to the architect handoff pattern.
